Update and rename groups/1506-2/kostjunin_av/1-test-version/solver.cpp to groups/1506-2/kostjunin_av/1-test-version/solver/solver.cpp
